+++ src/main/java/repositories/CheatsheetRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CheatsheetRepo extends JpaRepository<Cheatsheet, Integer> {
+public interface CheatsheetRepo extends JpaRepository<Cheatsheet, Long> {
 
 }
